Refactor prompt imports for better organization
Reorganize `COMMIT_PROMPT` and `SUMMARIZE_PROMPT` imports. Move `COMMIT_PROMPT` to `commands/commit/prompt`. Relocate `SUMMARIZE_PROMPT` to `langchain/chains/summarize/prompt`. This enhances code maintainability and module structure.
