+++ src/commands/init/handler.ts
@@ -3,7 +3,7 @@ import { Config } from '../types'
 import { appendToGitConfig } from '../../lib/config/services/git'
 import { appendToEnvFile } from '../../lib/config/services/env'
 import { logResult } from '../../lib/ui/logResult'
-import { COMMIT_PROMPT } from '../../lib/langchain/prompts/commit'
+import { COMMIT_PROMPT } from '../commit/prompt'
 import { appendToProjectJsonConfig } from '../../lib/config/services/project'
 import { LOGO } from '../../lib/ui/helpers'
 import { checkAndHandlePackageInstallation } from '../../lib/ui/checkAndHandlePackageInstall'
