+++ src/lib/langchain/chains/summarize/prompt.ts
